WIP: person-autocomplete
WIP: fixup "net_id" -- not just for netids

set ff=unix

reorganize so you can see what's going on better

'active' undefined reference cleaned up

reorganize lookup for nested function references

auto-select the first matching record on blur

remove redundant js requires

these are already required downstream and requiring them again is
causing errors

same for CSS

these CSS should be included by the parent app, (TODO: make sure this
information gets added to setup documentation)

add "ui-left-loading" class to place next to label

reindent some things

"Revealing module pattern"

and also, use the NDID instead of NetID since it's more reliably present
even for non-employees

turn everything into local vars that can be

some of these functions are still hoisted because reasons, also changing
these into "var" gives them more predictable semantics, so autocomplete
function is called to set up the autocomplete after the vars are defined

(hoisting a function ensures it does not inherit the local scope, this
is handy for example when you need "this" to point at the element that
received the event and not the module you're working inside of...)

expose ajax_params_base in personLookup

make it possible for an external function to use the personLookup module
to do your own person lookups, without making any autocomplete widgets

proof - this is the function

and actually, what we really wanted was...

not another global function that the personLookup consumer needs to know
about and learn to wield, actually... a function revealed by the module
(which the module itself will be calling on your behalf, saving you from
needing to learn about it when all you wanted was an autocomplete...)

but you could call it, if you wanted!

addClass will only work with the hoisted function

no idea what these were for

they don't make really any visible sense, unless the idea is to not add
a class without taking it away when the opposite action is taken...

JsonParser concern in the NdEmployeeLookup module

use_ssl is a named param and optional (default is true)
otherwise, this is identical to the JsonParser in Personnel Actions

WIP: permit an exact lookup

and update the search method to use the new JsonParser concern

add some convenience methods to Employee class

this was not working

Literally the "this" keyword was not working, because it should refer to
the select input and the function needs to be hoisted for that.
Established this pattern in an earlier commit.

don't offer "home orgn desc" which might be blank

add autoFocus for old employee_autocomplete.js

a bunch of these methods aren't working

they're trying to parse the JSON result from HrpyEmployeePerson.search,
which has already been parsed (and cannot be parsed again.)

also there has been "search" but no "find" method for too long, adding
one so every consumer does not have to implement their own exact_lookup

move api_error_message method to a common concern

add some new routes

WIP: controllers for new data lookup endpoints

without type: :controller, no render_views

spec dummy should not have included the migrations

not using capybara-webkit anymore

redo migrations with those changes

this test is broken because EmployeeJobLookup

there is no NdEmployeeJobLookup in the NdEmployeeLookup gem

fixup EmployeeStudentLookup and PersonLookup (red)

There are two new controllers and they are WIP, untested.  Need to
support "quick search" as in, lookup by indexed result and return a
listing for autocompletion... and "find" as in, exact lookup by ID only
(either NetID or NDID will do, for Person records that may have both.)

A third method might be handy here for finding those records by pidm,
but it may not be needed – if you have a pidm, then you already know the
person, you have completed a lookup before, and "employee-lookup-gem" is
not intended to be the utility providing you with answers.

(Never mind that it's called "employee lookup" and some of these records
you can use it to look up now, are not employees or students either.)

WIP: try implementing the new lookup controllers

add person-basic lookup endpoint to person lookup

since the endpoint exists and a service object is calling out to the
person API, placing this in the assets will allow it to be TDD'd now

fixup Person service model

it is important that search_url can have spaces

fixup: same for person_student_lookup.rb

format name, revised Person::V1::LookupController

the Person::V1::LookupController#search method now returns a limited set
of variables, basically Names, Search Name, NetID and NDID.

Format the name, including Middle Name if found, and format with NDID if
NetID is not found.

fixup for the right v_person_identity

searching for a person will yield a (usually ndid) or netid

(searching for an exact netid match will not usually yield an ndid)

this is the fastest way to build search so far, and it does not require
any joins at all!